move some files around (the routing changes will break the flow for now)
